requirements.txt
aiogram==3.*
apscheduler
python-dotenv